Transfer implementation of dynamic version checks to egg-info
